complete lesson_5

+++ geekshop/authapp/models.py
@@ -1,7 +1,7 @@
-from django.contrib.auth.models import AbstractUser
 from django.db import models
+from django.contrib.auth.models import AbstractUser
 
 
 class ShopUser(AbstractUser):
     avatar = models.ImageField(upload_to='users_avatars', blank=True)
-    age = models.PositiveSmallIntegerField(verbose_name='возраст')
+    age = models.PositiveIntegerField(verbose_name='возраст')
